+++ typescript/e2e/AppsIntranet/Tests/Custom/SetUpFixture.cs
@@ -2,11 +2,10 @@ namespace Tests
 {
     using System.Xml;
     using Allors.Database;
-    using Allors.Database.Adapters.Sql;
+    using Allors.Database.Adapters;
     using Allors.Database.Configuration;
     using Allors.Database.Domain;
     using NUnit.Framework;
-    using Database = Allors.Database.Adapters.Sql.SqlClient.Database;
     using Person = Allors.Database.Domain.Person;
 
     [SetUpFixture]
@@ -20,13 +19,10 @@ namespace Tests
 
             if (!Config.PopulationFileInfo.Exists)
             {
-                var database = new Database(
+                var database = new DatabaseBuilder(
                     new DefaultDatabaseServices(Config.Engine),
-                    new Configuration
-                    {
-                        ConnectionString = Config.Configuration["ConnectionStrings:DefaultConnection"],
-                        ObjectFactory = new ObjectFactory(Config.MetaPopulation, typeof(Person)),
-                    });
+                    Config.Configuration,
+                    new ObjectFactory(Config.MetaPopulation, typeof(Person))).Build();
 
                 database.Init();